requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.1
antlr4-python3-runtime==4.9.3
attrs==24.2.0
blessed==1.20.0
boto3==1.35.76
botocore==1.35.76
braceexpand==0.1.7
contourpy==1.3.1
cycler==0.12.1
einops==0.8.0
filelock==3.13.1
fonttools==4.55.2
frozenlist==1.5.0
fsspec==2024.2.0
gpustat==1.1.1
h5py==3.12.1
hydra-core==1.3.2
idna==3.10
Jinja2==3.1.3
jmespath==1.0.1
kiwisolver==1.4.7
lightning==2.4.0
lightning-utilities==0.11.9
MarkupSafe==2.1.5
matplotlib==3.9.3
mpmath==1.3.0
multidict==6.1.0
networkx==3.2.1
numpy==1.26.3
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-ml-py==12.560.30
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
omegaconf==2.3.0
packaging==24.2
pillow==10.2.0
propcache==0.2.1
protobuf==5.29.1
psutil==6.1.0
pyparsing==3.2.0
python-dateutil==2.9.0.post0
pytorch-lightning==2.4.0
PyYAML==6.0.2
s3transfer==0.10.4
scipy==1.14.1
setuptools==75.1.0
six==1.17.0
sympy==1.13.1
tensorboardX==2.6.2.2
tifffile==2024.9.20
torch==2.5.1+cu124
torchaudio==2.5.1+cu124
torchmetrics==1.6.0
torchvision==0.20.1+cu124
tqdm==4.67.1
triton==3.1.0
typing_extensions==4.9.0
urllib3==2.2.3
wcwidth==0.2.13
webdataset==0.2.100
wget==3.2
wheel==0.44.0
yarl==1.18.3
